+++ src/protocols/aggsig/mod.rs
@@ -18,4 +18,3 @@
 mod musig_two_rounds;
 mod musig_three_rounds;
 //mod test;
-
